Major update/refactoring
Somehow the UI code got messed up. I had to do a major refactoring on it to get everything to line up again. Logic wise, nothing should be affected.

Added the ability to use npm to install packages to the project

Added jshint and eslint packages to add javascript linting capabilities in VS Code and VS.

Ran an optimization script on the logo and reduced it's size by 50%! With no noticeable quality degradation.